backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-admin-interface==0.22.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-nonrelated-inlines==0.2
django-rest-swagger==2.2.0
django-seed==0.3.1
django-simple-history==3.2.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
drf-yasg2==1.19.4
Faker==15.3.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mollie-api-python==2.13.0
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.2
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
